feat(data-store): provide API for setting the database (data store) engine
Switch from choosing which DB engine to connect to based on database driver setting in .env file to
Providing an API for setting key/value pairs and use it to set the database (data store) engine.
This allows for more flexibility in working with databases (and also in setting and retrieving
values via keys using the package module)

BREAKING CHANGE: Apps relying on the previous method of setting and detecting the database engine
will break as the two methods differ completely both in philosophy and implementation. Also, though
minor, the directory "databases/drivers" has been replaced with "lib/stores".