fix(password validator): fix bug in password validation routine
When password validation data is set only via UserManager.config() call and not via environment
variables, the password validator does not see the data and returns null for the password
configuration object, and undefined for the member values of the object. When password validation is
set only via environment variables, the password validator also fails to pick up the values, instead
falling back to the default values. This commit fixes both issues.

"fix #2"